Remove Enforcer/Policy split
