+++ languages/js/src/index.ts
@@ -1,4 +1,3 @@
-export { Oso, Policy } from './Oso';
+export { Oso } from './Oso';
 export { Variable } from './Variable';
-export { Enforcer } from './Enforcer';
 export { AuthorizationError, ForbiddenError, NotFoundError } from './errors';
